refactor: create recordData utils
