+++ src/components/UpdateTemplate/UpdateTemplateContent.tsx
@@ -4,8 +4,8 @@ import Loading from '@/components/Common/Loading';
 import UpdateTemplateQuestionSelections from '@/components/UpdateTemplate/UpdateTemplateQuestionSelections';
 import UpdateTemplateSelectedQuestionContainer from '@/components/UpdateTemplate/UpdateTemplateSelectedQuestionContainer';
 import UpdateTemplateSubHeader from '@/components/UpdateTemplate/UpdateTemplateSubHeader';
-import useRecordDetail from '@/hooks/useRecordDetail';
 import { Questions, StringQuestionTypes } from '@/types/question.interface';
+import { useRecordDetail } from '@/utils/recordData';
 
 type TemplateContentProps = {
   setCurrTemplateSubHeader: Dispatch<
